requirements.txt
Django==1.11.7
Jinja2==2.9.6
MarkupSafe==1.0
Pillow==4.3.0
boto3==1.4.7
dj-database-url==0.4.2
django-model-utils==3.0.0
django-storages==1.6.5
django-widget-tweaks==1.4.1
gunicorn==19.7.1
psycopg2==2.7.3.2
requests
